requirements.txt
aiofiles==23.2.1
aiogram==2.22.2
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
pyaes==1.6.1
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
python-socks==2.4.4
pytz==2024.1
rsa==4.9
SQLAlchemy==2.0.27
Telethon==1.34.0
typing_extensions==4.9.0
yarl==1.9.4
